requirements.txt
cryptography
requests
werkzeug
Flask
Flask-SocketIO
Flask-Httpauth
gunicorn
pyOpenSSL
rsa
pycryptodomex